+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New command to check layers `yarn check-layers`
 - Vega Chart as component
 - Added Vega Chart Thumbnail with theme configuration
+- Added support for server rendering in Explore page (for search crawler)
 - Added support for GEE layers in Map
 - Pie charts display an "others" category to avoid duplicate colors